Included handlebars templating w/ inline template
I know I should've probably created another feature branch for this. I was bad and forgot to do so. There are other optimizations
and improvements that could be done, such as minifying the handlebars source, separating the template from content, returning all
fruits instead of one, etc. All to be done in due time!